% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorpusFromPDFDirectory.R
\name{getCorpusFromPDFDirectory}
\alias{getCorpusFromPDFDirectory}
\title{getCorpus}
\usage{
getCorpusFromPDFDirectory(literatureFileDirectory = "")
}
\arguments{
\item{literatureFileDirectory}{directory of pdf files to count words from.}
}
\value{
a corpus object containing the text of a directory of pdf files
}
\description{
This function accepts a folder of pdf journal articles and returns
a corpus object containing the text of the pdf files.
}
\examples{
getCorpus(literatureFileDirectory)


}
